+++ lib/sm2_response.dart
@@ -1,10 +1,10 @@
-/// Response type from calc.
-class SmResponse {
+/// Response type from the SM-2 calculation.
+class Sm2Response {
   final int interval;
   final int repetitions;
   final double easeFactor;
 
-  SmResponse({
+  Sm2Response({
     required this.interval,
     required this.repetitions,
     required this.easeFactor,